refactor(narwhals): rename phase-scaffolding tests to semantic module names
Replace phase-numbered test files with permanent, semantically-named modules:
- test_phase01_arch.py → test_error_handler.py
- test_arch03_schema_driven_dispatch.py → test_schema_driven_dtype_dispatch.py
- backends/test_e2e.py → backends/test_backend_parity.py
- test_phase02_validate_helper.py removed (covered by pyspark conftest)
